+++ src/Exception/InvalidMiddlewareException.php
@@ -7,6 +7,7 @@
 
 namespace Zend\Stratigility\Exception;
 
+use Interop\Http\Server\MiddlewareInterface;
 use InvalidArgumentException;
 
 class InvalidMiddlewareException extends InvalidArgumentException
@@ -19,15 +20,14 @@ class InvalidMiddlewareException extends InvalidArgumentException
      */
     public static function fromValue($value)
     {
-        $received = gettype($value);
-
-        if (is_object($value)) {
-            $received = get_class($value);
-        }
+        $received = is_object($value)
+            ? get_class($value)
+            : gettype($value);
 
         return new self(
             sprintf(
-                'Middleware must be callable, %s found',
+                'Middleware must implement %s; received middleware of type %s',
+                MiddlewareInterface::class,
                 $received
             )
         );