+++ src/Middleware/NotFoundHandler.php
@@ -7,13 +7,13 @@
 
 namespace Zend\Stratigility\Middleware;
 
+use Interop\Http\Server\MiddlewareInterface;
+use Interop\Http\Server\RequestHandlerInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
-use Webimpress\HttpMiddlewareCompatibility\HandlerInterface as DelegateInterface;
-use Webimpress\HttpMiddlewareCompatibility\MiddlewareInterface as ServerMiddlewareInterface;
 use Zend\Stratigility\Delegate\CallableDelegateDecorator;
 
-class NotFoundHandler implements ServerMiddlewareInterface
+class NotFoundHandler implements MiddlewareInterface
 {
     /**
      * @var ResponseInterface
@@ -49,10 +49,10 @@ class NotFoundHandler implements ServerMiddlewareInterface
      * Creates and returns a 404 response.
      *
      * @param ServerRequestInterface $request Ignored.
-     * @param DelegateInterface $delegate Ignored.
+     * @param RequestHandlerInterface $handler Ignored.
      * @return ResponseInterface
      */
-    public function process(ServerRequestInterface $request, DelegateInterface $delegate)
+    public function process(ServerRequestInterface $request, RequestHandlerInterface $handler) : ResponseInterface
     {
         $response = $this->responsePrototype
             ->withStatus(404);